% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sidra2.R
\name{get_sidra2}
\alias{get_sidra2}
\title{Get SIDRA's table}
\usage{
get_sidra2(x)
}
\arguments{
\item{x}{A character with the IBGE's SIDRA API parameters.}
}
\value{
The function returns a data frame printed by parameters input.
}
\description{
This function allows the user put directly the parameters of the query in the 
api format.
}
\seealso{
\code{\link{get_sidra}}
}
\author{
Renato Prado Siqueira \email{<rpradosiqueira@gmail.com>}
}
\keyword{IBGE}
\keyword{sidra}
