% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_sidra.R
\name{search_sidra}
\alias{search_sidra}
\title{Search SIDRA tables}
\usage{
search_sidra(x)
}
\arguments{
\item{x}{A non-empty character vector containing the search terms.}
}
\value{
A named character vector with matching SIDRA table titles. Names
are the table codes.
}
\description{
Searches table titles in the official IBGE aggregate catalog.
}
\examples{
\dontrun{
search_sidra("contas nacionais")
search_sidra("IPCA")
}
}
\seealso{
\code{\link[=get_sidra]{get_sidra()}} and \code{\link[=info_sidra]{info_sidra()}}
}
\author{
Renato Prado Siqueira \email{rpradosiqueira@gmail.com}
}
\keyword{IBGE}
\keyword{sidra}
