% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_sidra.R
\name{info_sidra}
\alias{info_sidra}
\title{Listing all the parameters of a SIDRA's table}
\usage{
info_sidra(x, wb = FALSE)
}
\arguments{
\item{x}{A table from SIDRA's API.}

\item{wb}{Logical. Should the metadata be open in the web browser?
Default to FALSE.}
}
\value{
A list with the all table's parameters.
}
\description{
It returns the parameters and their descriptions of a SIDRA's table.
}
\examples{
\dontrun{
info_sidra(1419)
}

}
\seealso{
\code{\link{get_sidra}}
}
\author{
Renato Prado Siqueira \email{rpradosiqueira@gmail.com}
}
\keyword{IBGE}
\keyword{sidra}
